Separate accessor generation visitors from operation generation visitors for Panache entities
Part of an effort to do this enhancement even when Panache is not in
use.
