+++ extensions/panache/hibernate-orm-panache-kotlin/deployment/src/main/java/io/quarkus/hibernate/orm/panache/kotlin/deployment/KotlinPanacheCompanionEnhancer.java
@@ -10,7 +10,7 @@ import org.objectweb.asm.ClassVisitor;
 
 import io.quarkus.panache.common.deployment.PanacheCompanionEnhancer;
 import io.quarkus.panache.common.deployment.PanacheMethodCustomizer;
-import io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassVisitor;
+import io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassOperationGenerationVisitor;
 
 public class KotlinPanacheCompanionEnhancer extends PanacheCompanionEnhancer {
 
@@ -20,7 +20,7 @@ public class KotlinPanacheCompanionEnhancer extends PanacheCompanionEnhancer {
 
     @Override
     public ClassVisitor apply(String className, ClassVisitor outputClassVisitor) {
-        return new KotlinPanacheClassVisitor(outputClassVisitor,
+        return new KotlinPanacheClassOperationGenerationVisitor(outputClassVisitor,
                 indexView.getClassByName(DotName.createSimple(className)), indexView, BUNDLE,
                 BUNDLE.entityCompanionBase(), methodCustomizers);
     }
