+++ extensions/panache/hibernate-orm-panache/deployment/src/main/java/io/quarkus/hibernate/orm/panache/deployment/PanacheJpaRepositoryEnhancer.java
@@ -5,7 +5,7 @@ import org.objectweb.asm.ClassVisitor;
 
 import io.quarkus.panache.common.deployment.PanacheRepositoryEnhancer;
 import io.quarkus.panache.common.deployment.TypeBundle;
-import io.quarkus.panache.common.deployment.visitors.PanacheRepositoryClassVisitor;
+import io.quarkus.panache.common.deployment.visitors.PanacheRepositoryClassOperationGenerationVisitor;
 
 public class PanacheJpaRepositoryEnhancer extends PanacheRepositoryEnhancer {
 
@@ -18,7 +18,7 @@ public class PanacheJpaRepositoryEnhancer extends PanacheRepositoryEnhancer {
 
     @Override
     public ClassVisitor apply(String className, ClassVisitor outputClassVisitor) {
-        return new PanacheRepositoryClassVisitor(className, outputClassVisitor,
+        return new PanacheRepositoryClassOperationGenerationVisitor(className, outputClassVisitor,
                 this.indexView, typeBundle);
     }
 }
