+++ extensions/panache/hibernate-reactive-panache/deployment/src/main/java/io/quarkus/hibernate/reactive/panache/deployment/PanacheJpaRepositoryEnhancer.java
@@ -4,7 +4,7 @@ import org.jboss.jandex.IndexView;
 import org.objectweb.asm.ClassVisitor;
 
 import io.quarkus.panache.common.deployment.PanacheRepositoryEnhancer;
-import io.quarkus.panache.common.deployment.visitors.PanacheRepositoryClassVisitor;
+import io.quarkus.panache.common.deployment.visitors.PanacheRepositoryClassOperationGenerationVisitor;
 
 public class PanacheJpaRepositoryEnhancer extends PanacheRepositoryEnhancer {
 
@@ -14,7 +14,7 @@ public class PanacheJpaRepositoryEnhancer extends PanacheRepositoryEnhancer {
 
     @Override
     public ClassVisitor apply(String className, ClassVisitor outputClassVisitor) {
-        return new PanacheRepositoryClassVisitor(className, outputClassVisitor,
+        return new PanacheRepositoryClassOperationGenerationVisitor(className, outputClassVisitor,
                 this.indexView, ReactiveJavaJpaTypeBundle.BUNDLE);
     }
 }
