+++ extensions/panache/mongodb-panache-kotlin/deployment/src/main/java/io/quarkus/mongodb/panache/kotlin/deployment/PanacheMongoCompanionEnhancer.java
@@ -11,7 +11,7 @@ import org.objectweb.asm.ClassVisitor;
 import io.quarkus.panache.common.deployment.PanacheCompanionEnhancer;
 import io.quarkus.panache.common.deployment.PanacheMethodCustomizer;
 import io.quarkus.panache.common.deployment.TypeBundle;
-import io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassVisitor;
+import io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassOperationGenerationVisitor;
 
 public class PanacheMongoCompanionEnhancer extends PanacheCompanionEnhancer {
     private final TypeBundle types;
@@ -24,7 +24,7 @@ public class PanacheMongoCompanionEnhancer extends PanacheCompanionEnhancer {
 
     @Override
     public ClassVisitor apply(String className, ClassVisitor outputClassVisitor) {
-        return new KotlinPanacheClassVisitor(outputClassVisitor,
+        return new KotlinPanacheClassOperationGenerationVisitor(outputClassVisitor,
                 indexView.getClassByName(DotName.createSimple(className)), indexView, types,
                 types.entityCompanionBase(), emptyList());
     }
