+++ extensions/panache/mongodb-panache/deployment/src/main/java/io/quarkus/mongodb/panache/deployment/PanacheMongoEntityEnhancer.java
@@ -1,24 +1,25 @@
 package io.quarkus.mongodb.panache.deployment;
 
 import static io.quarkus.mongodb.panache.deployment.BasePanacheMongoResourceProcessor.BSON_IGNORE;
-import static org.jboss.jandex.DotName.createSimple;
 
 import java.lang.reflect.Modifier;
 import java.util.List;
 
 import org.jboss.jandex.ClassInfo;
+import org.jboss.jandex.DotName;
 import org.jboss.jandex.FieldInfo;
 import org.jboss.jandex.IndexView;
 import org.objectweb.asm.ClassVisitor;
 
 import io.quarkus.gizmo.DescriptorUtils;
-import io.quarkus.mongodb.panache.deployment.visitors.PanacheMongoEntityClassVisitor;
+import io.quarkus.mongodb.panache.deployment.visitors.PanacheMongoEntityClassAccessorGenerationVisitor;
 import io.quarkus.panache.common.deployment.EntityField;
 import io.quarkus.panache.common.deployment.EntityModel;
 import io.quarkus.panache.common.deployment.MetamodelInfo;
 import io.quarkus.panache.common.deployment.PanacheEntityEnhancer;
 import io.quarkus.panache.common.deployment.PanacheMethodCustomizer;
 import io.quarkus.panache.common.deployment.TypeBundle;
+import io.quarkus.panache.common.deployment.visitors.PanacheEntityClassOperationGenerationVisitor;
 
 public class PanacheMongoEntityEnhancer extends PanacheEntityEnhancer {
 
@@ -33,8 +34,11 @@ public class PanacheMongoEntityEnhancer extends PanacheEntityEnhancer {
 
     @Override
     public ClassVisitor apply(String className, ClassVisitor outputClassVisitor) {
-        return new PanacheMongoEntityClassVisitor(outputClassVisitor, modelInfo,
-                indexView.getClassByName(createSimple(className)), methodCustomizers, typeBundle, indexView);
+        ClassInfo entityInfo = indexView.getClassByName(DotName.createSimple(className));
+        EntityModel entityModel = modelInfo.getEntityModel(className);
+        outputClassVisitor = new PanacheMongoEntityClassAccessorGenerationVisitor(outputClassVisitor, entityInfo, entityModel);
+        return new PanacheEntityClassOperationGenerationVisitor(outputClassVisitor, typeBundle,
+                entityInfo, methodCustomizers, indexView);
     }
 
     @Override
