+++ extensions/panache/panache-common/deployment/src/main/java/io/quarkus/panache/common/deployment/visitors/PanacheEntityClassOperationGenerationVisitor.java
@@ -1,14 +1,9 @@
 package io.quarkus.panache.common.deployment.visitors;
 
 import static io.quarkus.deployment.util.AsmUtil.getDescriptor;
-import static io.quarkus.panache.common.deployment.PanacheConstants.JAXB_ANNOTATION_PREFIX;
-import static io.quarkus.panache.common.deployment.PanacheConstants.JAXB_TRANSIENT_SIGNATURE;
-import static io.quarkus.panache.common.deployment.PanacheConstants.JSON_IGNORE_DOT_NAME;
-import static io.quarkus.panache.common.deployment.PanacheConstants.JSON_PROPERTY_DOT_NAME;
-import static io.quarkus.panache.common.deployment.PanacheConstants.JSON_PROPERTY_SIGNATURE;
-import static io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassVisitor.OBJECT;
-import static io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassVisitor.recursivelyFindEntityTypeArguments;
-import static io.quarkus.panache.common.deployment.visitors.PanacheRepositoryClassVisitor.CLASS;
+import static io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassOperationGenerationVisitor.OBJECT;
+import static io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassOperationGenerationVisitor.recursivelyFindEntityTypeArguments;
+import static io.quarkus.panache.common.deployment.visitors.PanacheRepositoryClassOperationGenerationVisitor.CLASS;
 import static org.objectweb.asm.Opcodes.CHECKCAST;
 import static org.objectweb.asm.Opcodes.INVOKEVIRTUAL;
 
@@ -25,13 +20,10 @@ import java.util.function.Function;
 import org.jboss.jandex.AnnotationInstance;
 import org.jboss.jandex.AnnotationValue;
 import org.jboss.jandex.ClassInfo;
-import org.jboss.jandex.FieldInfo;
 import org.jboss.jandex.IndexView;
 import org.jboss.jandex.MethodInfo;
 import org.jboss.jandex.TypeVariable;
-import org.objectweb.asm.AnnotationVisitor;
 import org.objectweb.asm.ClassVisitor;
-import org.objectweb.asm.FieldVisitor;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
@@ -39,20 +31,17 @@ import org.objectweb.asm.Type;
 import io.quarkus.deployment.util.AsmUtil;
 import io.quarkus.gizmo.Gizmo;
 import io.quarkus.panache.common.deployment.ByteCodeType;
-import io.quarkus.panache.common.deployment.EntityField;
-import io.quarkus.panache.common.deployment.EntityField.EntityFieldAnnotation;
-import io.quarkus.panache.common.deployment.EntityModel;
-import io.quarkus.panache.common.deployment.MetamodelInfo;
 import io.quarkus.panache.common.deployment.PanacheConstants;
 import io.quarkus.panache.common.deployment.PanacheMethodCustomizer;
 import io.quarkus.panache.common.deployment.PanacheMethodCustomizerVisitor;
-import io.quarkus.panache.common.deployment.PanacheMovingAnnotationVisitor;
 import io.quarkus.panache.common.deployment.TypeBundle;
 
-public abstract class PanacheEntityClassVisitor extends ClassVisitor {
+/**
+ * A visitor that adds Panache operations to a Panache entity type.
+ */
+public class PanacheEntityClassOperationGenerationVisitor extends ClassVisitor {
 
     protected Type thisClass;
-    protected final Map<String, ? extends EntityField> fields;
     private final Set<String> userMethods = new HashSet<>();
     protected TypeBundle typeBundle;
     protected final ClassInfo panacheEntityBaseClassInfo;
@@ -63,8 +52,7 @@ public abstract class PanacheEntityClassVisitor extends ClassVisitor {
     protected final ByteCodeType entityUpperBound;
     private final Map<String, String> erasures = new HashMap<>();
 
-    public PanacheEntityClassVisitor(ClassVisitor outputClassVisitor,
-            MetamodelInfo modelInfo,
+    public PanacheEntityClassOperationGenerationVisitor(ClassVisitor outputClassVisitor,
             TypeBundle typeBundle,
             ClassInfo entityInfo,
             List<PanacheMethodCustomizer> methodCustomizers, IndexView indexView) {
@@ -73,8 +61,6 @@ public abstract class PanacheEntityClassVisitor extends ClassVisitor {
         String className = entityInfo.name().toString();
         thisClass = Type.getType("L" + className.replace('.', '/') + ";");
         this.typeBundle = typeBundle;
-        EntityModel entityModel = modelInfo.getEntityModel(className);
-        fields = entityModel != null ? entityModel.fields : null;
         this.panacheEntityBaseClassInfo = indexView.getClassByName(typeBundle.entityBase().dotName());
         this.entityInfo = entityInfo;
         this.methodCustomizers = methodCustomizers;
@@ -93,53 +79,6 @@ public abstract class PanacheEntityClassVisitor extends ClassVisitor {
             return byteCodeType != null
                     ? byteCodeType.descriptor()
                     : OBJECT.descriptor();
-        };
-    }
-
-    @Override
-    public FieldVisitor visitField(int access, String name, String descriptor, String signature, Object value) {
-        if (fields == null || fields.isEmpty()) {
-            return super.visitField(access, name, descriptor, signature, value);
-        }
-        EntityField ef = fields.get(name);
-        if (ef == null) {
-            return super.visitField(access, name, descriptor, signature, value);
-        }
-        //we make the fields protected
-        //so any errors are visible immediately, rather than data just being lost
-
-        FieldVisitor superVisitor;
-        if (name.equals("id")) {
-            superVisitor = super.visitField(access, name, descriptor, signature, value);
-        } else {
-            superVisitor = super.visitField((access | Modifier.PROTECTED) & ~(Modifier.PRIVATE | Modifier.PUBLIC),
-                    name, descriptor, signature, value);
-        }
-        ef.signature = signature;
-        // if we have a mapped field, let's add some annotations
-        return new FieldVisitor(Gizmo.ASM_API_VERSION, superVisitor) {
-            @Override
-            public AnnotationVisitor visitAnnotation(String descriptor, boolean visible) {
-                if (!descriptor.startsWith(JAXB_ANNOTATION_PREFIX)) {
-                    return super.visitAnnotation(descriptor, visible);
-                } else {
-                    // Save off JAX-B annotations on the field so they can be applied to the generated getter later
-                    EntityFieldAnnotation efAnno = new EntityFieldAnnotation(descriptor);
-                    ef.annotations.add(efAnno);
-                    return new PanacheMovingAnnotationVisitor(efAnno);
-                }
-            }
-
-            @Override
-            public void visitEnd() {
-                // Add the @JaxbTransient property to the field so that JAXB prefers the generated getter (otherwise JAXB complains about
-                // having a field and property both with the same name)
-                // JSONB will already use the getter so we're good
-                // Note: we don't need to check if we already have @XmlTransient in the descriptors because if we did, we moved it to the getter
-                // so we can't have any duplicate
-                super.visitAnnotation(JAXB_TRANSIENT_SIGNATURE, true);
-                super.visitEnd();
-            }
         };
     }
 
@@ -178,8 +117,6 @@ public abstract class PanacheEntityClassVisitor extends ClassVisitor {
                 }
             }
         }
-
-        generateAccessors();
 
         super.visitEnd();
     }
@@ -305,102 +242,4 @@ public abstract class PanacheEntityClassVisitor extends ClassVisitor {
         mv.visitLdcInsn(thisClass);
     }
 
-    protected void generateAccessors() {
-        if (fields == null)
-            return;
-        for (EntityField field : fields.values()) {
-            // Getter
-            String getterName = field.getGetterName();
-            String getterDescriptor = "()" + field.descriptor;
-            if (!userMethods.contains(getterName + "/" + getterDescriptor)) {
-                MethodVisitor mv = super.visitMethod(Opcodes.ACC_PUBLIC,
-                        getterName, getterDescriptor, field.signature == null ? null : "()" + field.signature, null);
-                mv.visitCode();
-                mv.visitIntInsn(Opcodes.ALOAD, 0);
-                generateAccessorGetField(mv, field);
-                int returnCode = AsmUtil.getReturnInstruction(field.descriptor);
-                mv.visitInsn(returnCode);
-                mv.visitMaxs(0, 0);
-                // Apply JAX-B annotations that are being transferred from the field
-                for (EntityFieldAnnotation anno : field.annotations) {
-                    anno.writeToVisitor(mv);
-                }
-                // Add an explicit Jackson annotation so that the entire property is not ignored due to having @XmlTransient
-                // on the field
-                if (shouldAddJsonProperty(field)) {
-                    AnnotationVisitor visitor = mv.visitAnnotation(JSON_PROPERTY_SIGNATURE, true);
-                    FieldInfo fieldInfo = entityInfo.field(field.name);
-                    if (fieldInfo != null) {
-                        AnnotationInstance jsonPropertyInstance = fieldInfo.annotation(JSON_PROPERTY_DOT_NAME);
-                        // propagate the value of @JsonProperty field annotation to the newly added method annotation
-                        if (jsonPropertyInstance != null) {
-                            AnnotationValue jsonPropertyValue = jsonPropertyInstance.value();
-                            if ((jsonPropertyValue != null) && !jsonPropertyValue.asString().isEmpty()) {
-                                visitor.visit("value", jsonPropertyValue.asString());
-                            }
-                        }
-                    }
-                    visitor.visitEnd();
-                }
-                mv.visitEnd();
-            }
-
-            // Setter
-            String setterName = field.getSetterName();
-            String setterDescriptor = "(" + field.descriptor + ")V";
-            if (!userMethods.contains(setterName + "/" + setterDescriptor)) {
-                MethodVisitor mv = super.visitMethod(Opcodes.ACC_PUBLIC,
-                        setterName, setterDescriptor, field.signature == null ? null : "(" + field.signature + ")V", null);
-                mv.visitCode();
-                mv.visitIntInsn(Opcodes.ALOAD, 0);
-                int loadCode;
-                switch (field.descriptor) {
-                    case "Z":
-                    case "B":
-                    case "C":
-                    case "S":
-                    case "I":
-                        loadCode = Opcodes.ILOAD;
-                        break;
-                    case "J":
-                        loadCode = Opcodes.LLOAD;
-                        break;
-                    case "F":
-                        loadCode = Opcodes.FLOAD;
-                        break;
-                    case "D":
-                        loadCode = Opcodes.DLOAD;
-                        break;
-                    default:
-                        loadCode = Opcodes.ALOAD;
-                        break;
-                }
-                mv.visitIntInsn(loadCode, 1);
-                generateAccessorSetField(mv, field);
-                mv.visitInsn(Opcodes.RETURN);
-                mv.visitMaxs(0, 0);
-                mv.visitEnd();
-            }
-        }
-    }
-
-    private boolean shouldAddJsonProperty(EntityField entityField) {
-        if (isAnnotatedWithJsonIgnore(entityField)) {
-            return false;
-        }
-        return !entityField.hasAnnotation(JSON_PROPERTY_SIGNATURE);
-    }
-
-    private boolean isAnnotatedWithJsonIgnore(EntityField entityField) {
-        FieldInfo field = entityInfo.field(entityField.name);
-        if (field != null) {
-            return field.hasAnnotation(JSON_IGNORE_DOT_NAME);
-        }
-
-        return false;
-    }
-
-    protected abstract void generateAccessorSetField(MethodVisitor mv, EntityField field);
-
-    protected abstract void generateAccessorGetField(MethodVisitor mv, EntityField field);
 }
