+++ extensions/panache/panache-common/deployment/src/main/java/io/quarkus/panache/common/deployment/visitors/PanacheRepositoryClassOperationGenerationVisitor.java
@@ -2,8 +2,8 @@ package io.quarkus.panache.common.deployment.visitors;
 
 import static io.quarkus.deployment.util.AsmUtil.getDescriptor;
 import static io.quarkus.deployment.util.AsmUtil.unboxIfRequired;
-import static io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassVisitor.OBJECT;
-import static io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassVisitor.recursivelyFindEntityTypeArguments;
+import static io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassOperationGenerationVisitor.OBJECT;
+import static io.quarkus.panache.common.deployment.visitors.KotlinPanacheClassOperationGenerationVisitor.recursivelyFindEntityTypeArguments;
 import static org.objectweb.asm.Opcodes.CHECKCAST;
 import static org.objectweb.asm.Opcodes.INVOKEVIRTUAL;
 
@@ -32,7 +32,7 @@ import io.quarkus.panache.common.deployment.ByteCodeType;
 import io.quarkus.panache.common.deployment.PanacheConstants;
 import io.quarkus.panache.common.deployment.TypeBundle;
 
-public class PanacheRepositoryClassVisitor extends ClassVisitor {
+public class PanacheRepositoryClassOperationGenerationVisitor extends ClassVisitor {
     public static final ByteCodeType CLASS = new ByteCodeType(Class.class);
 
     protected Type entityType;
@@ -49,7 +49,8 @@ public class PanacheRepositoryClassVisitor extends ClassVisitor {
     protected ByteCodeType entityUpperBound;
     private final Map<String, String> erasures = new HashMap<>();
 
-    public PanacheRepositoryClassVisitor(String className, ClassVisitor outputClassVisitor, IndexView indexView,
+    public PanacheRepositoryClassOperationGenerationVisitor(String className, ClassVisitor outputClassVisitor,
+            IndexView indexView,
             TypeBundle typeBundle) {
         super(Gizmo.ASM_API_VERSION, outputClassVisitor);
         this.typeBundle = typeBundle;
