+++ extensions/panache/panache-hibernate-common/deployment/src/main/java/io/quarkus/panache/common/deployment/visitors/PanacheJpaEntityClassAccessorGenerationVisitor.java
@@ -1,26 +1,20 @@
 package io.quarkus.panache.common.deployment.visitors;
 
-import java.util.List;
-
 import org.hibernate.bytecode.enhance.spi.EnhancerConstants;
 import org.jboss.jandex.ClassInfo;
-import org.jboss.jandex.IndexView;
 import org.objectweb.asm.ClassVisitor;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
 
 import io.quarkus.panache.common.deployment.EntityField;
-import io.quarkus.panache.common.deployment.MetamodelInfo;
-import io.quarkus.panache.common.deployment.PanacheMethodCustomizer;
-import io.quarkus.panache.common.deployment.TypeBundle;
+import io.quarkus.panache.common.deployment.EntityModel;
 
-public class PanacheJpaEntityClassVisitor extends PanacheEntityClassVisitor {
+public class PanacheJpaEntityClassAccessorGenerationVisitor extends PanacheEntityClassAccessorGenerationVisitor {
 
-    public PanacheJpaEntityClassVisitor(ClassVisitor outputClassVisitor,
-            MetamodelInfo modelInfo, ClassInfo entityInfo,
-            List<PanacheMethodCustomizer> methodCustomizers, IndexView indexView, TypeBundle bundle) {
-        super(outputClassVisitor, modelInfo, bundle, entityInfo, methodCustomizers, indexView);
+    public PanacheJpaEntityClassAccessorGenerationVisitor(ClassVisitor outputClassVisitor,
+            ClassInfo entityInfo, EntityModel entityModel) {
+        super(outputClassVisitor, entityInfo, entityModel);
     }
 
     @Override
